Change to unsigned long for hits and misses.

+++ run_stats_types.h
@@ -78,8 +78,8 @@ public:
     unsigned long int m_bytes_rx;
     unsigned long int m_bytes_tx;
     unsigned long int m_ops;
-    unsigned int m_hits;
-    unsigned int m_misses;
+    unsigned long int m_hits;
+    unsigned long int m_misses;
     unsigned int m_moved;
     unsigned int m_ask;
     unsigned long long int m_total_latency;
